refactor: atom 상태관리 hook으로 추출
